Move arranging testers into their own namespace
